gender-predictor-69/requirements.txt
gunicorn
flask
keras
numpy
tensorflow
Pillow
requests
